requirements.txt
boto3
pyyaml
flask-ask
zappa
